+layout.svelte (NEU) — Shared Layout für alle /cardsboard/-Routen. Enthält Topbar, Left Sidebar (BoardsList), Right Sidebar (AI), Board-Rendering, Mobile-Detection, Comment-Subscriptions. Child-Pages rendern nur Overlays/Dialogs via {@render children()}.
+page.svelte (vereinfacht, von ~410 auf ~90 Zeilen) — Nur noch Import-Link-Handling (?import=, ?share=), FollowBoardDialog und RequestEditorRoleDialog. Das Board-UI kommt komplett vom Layout.

[naddr]/+page.svelte (angepasst) — Kein goto() Redirect mehr. Stattdessen:

Während des Ladens: Fixed-Overlay (z-50) über dem Board
Nach dem Laden: Overlay verschwindet, Board wird vom Layout gerendert
URL bleibt als permanenter Share-Link erhalten: /cardsboard/naddr1qvz...
